LayoutTests/svg/dynamic-updates/SVGFESpotLightElement-svgdom-pointsAtY-prop-expected.txt

CONSOLE MESSAGE: line 44: TypeError: 'undefined' is not an object (evaluating 'spotLightElement.pointsAtY.baseVal = 100')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'pointsAtY' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL spotLightElement.pointsAtY.baseVal should be 0. Threw exception TypeError: 'undefined' is not an object (evaluating 'spotLightElement.pointsAtY.baseVal')
